PR 5: Decompose LibraryScreen monolith (#906)
Extracted composables from 1033-line LibraryScreen.kt into focused files:

**LibraryMangaGrid.kt** — MangaGrid, UnreadBadge, isManhwa helper
- Supports both LazyVerticalGrid and LazyVerticalStaggeredGrid
- Content-type filter tabs (All / Manga / Manhwa) with derivedStateOf
- Animation support via AnimatedVisibility + scaleIn

**LibraryEmptyState.kt** — EmptyLibraryMessage + @Preview

**LibraryCategoryFilter.kt** — CategoryFilterChips, ReadingListFilterChips

**LibraryContinueReading.kt** — ContinueReadingSection, ContinueReadingCard

**LibraryMangaDetailPanel.kt** — MangaDetailPanel for expanded-width layout

**LibraryDailyGoal.kt** — DailyGoalBanner with LinearProgressIndicator

LibraryScreen.kt now ~270 lines (route + LibraryContent only).

Closes #906
