+++ feature/library/src/main/java/app/otakureader/feature/library/LibraryDailyGoal.kt
@@ -1,0 +1,102 @@
+package app.otakureader.feature.library
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.res.pluralStringResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.foundation.shape.RoundedCornerShape
+import app.otakureader.domain.model.ReadingGoal
+
+/**
+ * Compact banner showing progress toward today's chapter reading goal.
+ * Displayed at the top of the library grid whenever a daily goal is configured.
+ */
+@Composable
+internal fun DailyGoalBanner(
+    readingGoal: ReadingGoal,
+    modifier: Modifier = Modifier
+) {
+    // Guard: dailyGoal must be positive before computing the fraction to avoid division by zero.
+    if (readingGoal.dailyGoal <= 0) return
+    val isComplete = readingGoal.dailyProgress >= readingGoal.dailyGoal
+    val fraction = (readingGoal.dailyProgress.toFloat() / readingGoal.dailyGoal.toFloat())
+        .coerceIn(0f, 1f)
+
+    Card(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(vertical = 4.dp),
+        colors = CardDefaults.cardColors(
+            containerColor = if (isComplete)
+                MaterialTheme.colorScheme.primaryContainer
+            else
+                MaterialTheme.colorScheme.surfaceVariant
+        ),
+        elevation = CardDefaults.cardElevation(defaultElevation = 0.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 12.dp, vertical = 8.dp),
+            verticalArrangement = Arrangement.spacedBy(4.dp)
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = stringResource(R.string.library_daily_goal_label),
+                    style = MaterialTheme.typography.labelMedium,
+                    fontWeight = FontWeight.SemiBold,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Text(
+                    text = if (isComplete)
+                        stringResource(R.string.library_daily_goal_complete)
+                    else
+                        pluralStringResource(
+                            R.plurals.library_daily_goal_progress,
+                            readingGoal.dailyGoal,
+                            readingGoal.dailyProgress,
+                            readingGoal.dailyGoal
+                        ),
+                    style = MaterialTheme.typography.labelMedium,
+                    color = if (isComplete)
+                        MaterialTheme.colorScheme.primary
+                    else
+                        MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+            LinearProgressIndicator(
+                progress = { fraction },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(6.dp)
+                    .clip(RoundedCornerShape(3.dp)),
+                color = if (isComplete)
+                    MaterialTheme.colorScheme.primary
+                else
+                    MaterialTheme.colorScheme.secondary,
+                trackColor = MaterialTheme.colorScheme.surface,
+                strokeCap = StrokeCap.Round,
+            )
+        }
+    }
+}
